Move all config files to config package
